Fix ESP32-C3 firmware build: format specifiers and linker ordering
Cast int32_t to (int) for %d format specifiers in rb_array.c and
rb_input.c to fix -Werror=format on RISC-V where int32_t is long int.
Add WHOLE_ARCHIVE to idf_component_register so runtime symbols remain
available for basic_program.o during linking.

+++ runtime/src/rb_array.c
@@ -7,14 +7,14 @@
 void rb_array_check_dim_size(int32_t dim_value, int32_t dim_index) {
     if (dim_value < 0) {
         fprintf(stderr, "Negative array dimension size: DIM dimension %d = %d\n",
-                dim_index, dim_value);
+                (int)dim_index, (int)dim_value);
         rb_panic("negative array dimension size");
     }
 }
 
 void* rb_array_alloc(int32_t element_size, int32_t total_elements) {
     if (total_elements <= 0) {
-        fprintf(stderr, "Invalid array total size: %d\n", total_elements);
+        fprintf(stderr, "Invalid array total size: %d\n", (int)total_elements);
         rb_panic("invalid array size");
     }
     if (element_size <= 0) {
@@ -23,7 +23,7 @@ void* rb_array_alloc(int32_t element_size, int32_t total_elements) {
     /* Overflow check: element_size * total_elements must fit in size_t */
     if ((size_t)total_elements > SIZE_MAX / (size_t)element_size) {
         fprintf(stderr, "Array allocation overflow: %d * %d elements\n",
-                element_size, total_elements);
+                (int)element_size, (int)total_elements);
         rb_panic("array allocation size overflow");
     }
     size_t bytes = (size_t)element_size * (size_t)total_elements;
@@ -42,7 +42,7 @@ void rb_array_free(void* ptr) {
 
 void rb_array_bounds_check(int32_t index, int32_t size) {
     if (index < 0 || index >= size) {
-        fprintf(stderr, "Array index out of bounds: index %d, size %d\n", index, size);
+        fprintf(stderr, "Array index out of bounds: index %d, size %d\n", (int)index, (int)size);
         rb_panic("array index out of bounds");
     }
 }
